fix: manually rework certain form names
Some species on PokeAPI have a default form with a non-null form name.

This on occasion doesn't make sense for the matchup tool, e.g. there's
no need to show Deoxys as Deoxys (Normal), when we only show that single
form because the others are all the same type.

So, I went through and adjusted the first form names for a few species,
both in the pokemon list and on their corresponding sprite file names.